+++ internal/envctx/envctx.go
@@ -17,6 +17,8 @@ type Context struct {
 	KeyFile  string
 	Vars     map[string]string
 	VarOrder []string
+	Allow    []string
+	AllowSet bool
 	Pointer  *pointer.Pointer
 }
 
@@ -40,6 +42,7 @@ func Resolve(cliEnv, startDir string) (*Context, error) {
 		Env: env, Source: source,
 		Vault: ep.Vault, KeyFile: ep.KeyFile,
 		Vars: ep.Vars, VarOrder: ep.VarOrder,
+		Allow: ep.Allow, AllowSet: ep.AllowSet,
 		Pointer: p,
 	}, nil
 }